refactor: remove task validator

+++ src/main/java/dev/matheuslf/desafio/inscritos/service/TaskService.java
@@ -11,12 +11,12 @@ import dev.matheuslf.desafio.inscritos.entities.enums.Priority;
 import dev.matheuslf.desafio.inscritos.entities.enums.Role;
 import dev.matheuslf.desafio.inscritos.entities.enums.Status;
 import dev.matheuslf.desafio.inscritos.exception.BusinessException;
+import dev.matheuslf.desafio.inscritos.exception.ConflictException;
 import dev.matheuslf.desafio.inscritos.exception.ResourceNotFoundException;
 import dev.matheuslf.desafio.inscritos.mapper.TaskMapper;
 import dev.matheuslf.desafio.inscritos.repository.ProjectRepository;
 import dev.matheuslf.desafio.inscritos.repository.TaskRepository;
 import dev.matheuslf.desafio.inscritos.repository.UserRepository;
-import dev.matheuslf.desafio.inscritos.validator.TaskValidator;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
@@ -26,6 +26,7 @@ import org.springframework.stereotype.Service;
 
 import java.time.LocalDate;
 import java.util.List;
+import java.util.Optional;
 import java.util.UUID;
 
 import static dev.matheuslf.desafio.inscritos.repository.specs.TaskSpec.*;
@@ -36,7 +37,6 @@ public class TaskService {
 
     private static final String TASK_NOT_FOUND_MESSAGE = "Task not found with id: ";
     private final TaskRepository taskRepository;
-    private final TaskValidator taskValidator;
     private final TaskMapper taskMapper;
     private final ProjectRepository projectRepository;
     private final UserRepository userRepository;
@@ -44,7 +44,9 @@ public class TaskService {
     public TaskResponseDTO save(TaskRequestDTO dto) {
         Task task = taskMapper.toEntity(dto);
 
-        taskValidator.validateTaskName(task);
+        if (existsRegisteredTask(task)) {
+            throw new ConflictException("There is already a task with this name in this project");
+        }
 
         if (task.getProject().getEndDate().isBefore(LocalDate.now())) {
             throw new BusinessException("Project has already ended");
@@ -57,13 +59,16 @@ public class TaskService {
     }
 
     public TaskResponseDTO updateStatus(User user, UUID id, TaskStatusUpdateDTO dto) {
-        Task task = getTaskById(id);
+        Task task = taskRepository.findById(id)
+                .orElseThrow( () -> new ResourceNotFoundException(TASK_NOT_FOUND_MESSAGE + id));
 
         if (!(user.getRole().equals(Role.ADMIN) || task.getAssignee().getEmail().equals(user.getEmail()))) {
             throw new BusinessException("You do not have permission to update the status of this task");
         }
 
-        taskValidator.validateTaskStatus(task);
+        if (task.getStatus().equals(Status.DONE)) {
+            throw new ConflictException("It is not possible to modify an completed task");
+        }
 
         task.setStatus(Status.valueOf(dto.status()));
         Task savedTask = taskRepository.save(task);
@@ -105,14 +110,20 @@ public class TaskService {
     }
 
     public void delete(UUID id) {
-        Task task = getTaskById(id);
+        Task task = taskRepository.findById(id)
+                .orElseThrow( () -> new ResourceNotFoundException(TASK_NOT_FOUND_MESSAGE + id));
 
         if (task.getProject().getEndDate().isBefore(LocalDate.now())) {
             throw new BusinessException("Project has already ended");
         }
 
-        taskValidator.validateTaskDueDate(task);
-        taskValidator.validateTaskStatus(task);
+        if (task.getDueDate().isAfter(LocalDate.now())) {
+            throw new ConflictException("It is not possible to modify an expired task");
+        }
+
+        if (task.getStatus().equals(Status.DONE)) {
+            throw new ConflictException("It is not possible to modify an completed task");
+        }
 
         taskRepository.delete(task);
     }
@@ -120,26 +131,35 @@ public class TaskService {
     public List<TaskResponseDTO> findByProject(UUID projectId) {
         Project project = projectRepository.findById(projectId)
                 .orElseThrow( () -> new ResourceNotFoundException("Project not found with id: " + projectId));
+
         return taskRepository.findAllByProject(project).stream()
                 .map(taskMapper::toDTO)
                 .toList();
     }
 
     public TaskResponseDTO findById(UUID id) {
-        Task task = getTaskById(id);
+        Task task = taskRepository.findById(id)
+                .orElseThrow( () -> new ResourceNotFoundException(TASK_NOT_FOUND_MESSAGE + id));
+
         return taskMapper.toDTO(task);
     }
 
     public List<TaskResponseDTO> findByAssignee(UUID assigneeId) {
         User assignee = userRepository.findById(assigneeId)
                 .orElseThrow( () -> new ResourceNotFoundException("Assignee not found with id: " + assigneeId));
+
         return taskRepository.findAllByAssignee(assignee).stream()
                 .map(taskMapper::toDTO)
                 .toList();
     }
 
-    private Task getTaskById(UUID id) {
-        return taskRepository.findById(id)
-                .orElseThrow( () -> new ResourceNotFoundException(TASK_NOT_FOUND_MESSAGE + id));
+    private boolean existsRegisteredTask(Task task) {
+        Optional<Task> foundTask = taskRepository.findByTitleAndProject(task.getTitle(), task.getProject());
+
+        if (task.getId() == null) {
+            return foundTask.isPresent();
+        }
+
+        return foundTask.isPresent() && !task.getId().equals(foundTask.get().getId());
     }
 }
